To factoring, add new methods to set and get raw decoding settings. Mark all sub methods are depreciated. There will be removed for KDE 4.4.2.
svn path=/trunk/KDE/kdegraphics/libs/libkdcraw/; revision=1108593

+++ libkdcraw/rawdecodingsettings.h
@@ -6,8 +6,8 @@
  * Date        : 2006-12-09
  * Description : Raw decoding settings
  *
- * Copyright (C) 2006-2009 by Gilles Caulier <caulier dot gilles at gmail dot com>
- * Copyright (C) 2006-2009 by Marcel Wiesweg <marcel dot wiesweg at gmx dot de>
+ * Copyright (C) 2006-2010 by Gilles Caulier <caulier dot gilles at gmail dot com>
+ * Copyright (C) 2006-2010 by Marcel Wiesweg <marcel dot wiesweg at gmx dot de>
  * Copyright (C) 2007-2008 by Guillaume Castagnino <casta at xwing dot info>
  *
  * This program is free software; you can redistribute it
@@ -203,7 +203,7 @@ public:
 
     /** Brightness of output image.
     */
-    float brightness;
+    double brightness;
 
     /** Set on the black point setting to decode RAW image.
     */
